Consolidate and abstract spreadsheet services; add method for unavailable specialists.
